+++ lib/part-structs-array.c
@@ -3,7 +3,7 @@
 #define INITIAL_SIZE 30
 
 struct parts_type {
-  int count;
+  size_t count;
   struct part_type *rows;
   size_t requested_row_allocation;
 };
@@ -28,7 +28,7 @@ void destroy_db(Parts db)
   free(db->rows);
   free(db);
 }
-int size(Parts db)
+size_t size(Parts db)
 {
   return db->count;
 }
@@ -44,7 +44,7 @@ static int resize_db(Parts db)
 }
 int delete_part(Parts db, PartNumber number)
 {
-  int i;
+  size_t i;
 
   for (i = 0; i < db->count; i++)
     if (get_part_number(&db->rows[i]) == number)
@@ -62,9 +62,9 @@ int delete_part(Parts db, PartNumber number)
 }
 int insert_part(Parts db, Part p)
 {
-  int i, j;
+  size_t i, j;
 
-  if (db->count == (int)db->requested_row_allocation)
+  if (db->count == db->requested_row_allocation)
     if ((resize_db(db) != 0)) {
       return -1;
     }
@@ -93,7 +93,7 @@ int insert_part(Parts db, Part p)
  **********************************************************/
 Part find_part(Parts db, PartNumber part_number)
 {
-  int i;
+  size_t i;
   for (i = 0; i < db->count; i++)
     if (get_part_number(&db->rows[i]) == part_number)
       return &db->rows[i];
@@ -101,7 +101,7 @@ Part find_part(Parts db, PartNumber part_number)
 }
 void iterate(Parts db, void (*op)(Part p))
 {
-  for (int i = 0; i < db->count; i++)
+  for (size_t i = 0; i < db->count; i++)
     op(&db->rows[i]);
 }
 
@@ -113,7 +113,7 @@ int dump(char *outfile, Parts db)
     return -1;
 
   n_written = fwrite(db->rows, sizeof(db->rows[0]), db->count, ostream);
-  if (n_written < (size_t) db->count || ferror(ostream))
+  if (n_written < db->count || ferror(ostream))
     return -2;
 
   if (fclose(ostream) == EOF)
@@ -141,7 +141,7 @@ Parts restore(char *infile)
         break;
     }
     db->count += n_read;
-    if ((size_t)db->count >= db->requested_row_allocation) {
+    if (db->count >= db->requested_row_allocation) {
       if (resize_db(db) != 0) {
         destroy_db(db);
         return NULL;
